requirements.txt
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
boto3==1.18.54
botocore==1.21.54
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
Django==3.2.6
django-cors-headers==3.8.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.20.1
drf-spectacular-sidecar==2021.9.27
idna==3.2
inflection==0.5.1
ipython==7.27.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==4.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mysqlclient==2.0.3
packaging==21.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.1.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.1.0
uritemplate==3.0.1
urllib3==1.26.7
wcwidth==0.2.5
